+++ minijinja-cli/tests/test_basic.rs
@@ -12,13 +12,13 @@ fn file_with_contents(contents: &str) -> NamedTempFile {
     file_with_contents_and_ext(contents, "")
 }
 
-fn file_with_contents_and_ext(contents: &str, ext: &str) -> NamedTempFile {
+fn file_with_contents_and_ext<X: AsRef<[u8]>>(contents: X, ext: &str) -> NamedTempFile {
     let mut f = tempfile::Builder::new()
         .prefix("minijinja-testfile--")
         .suffix(ext)
         .tempfile()
         .unwrap();
-    f.write_all(contents.as_bytes()).unwrap();
+    f.write_all(contents.as_ref()).unwrap();
     f
 }
 
@@ -233,6 +233,29 @@ fn test_querystring() {
 }
 
 #[test]
+#[cfg(feature = "cbor")]
+fn test_cbor() {
+    let input = file_with_contents_and_ext(
+        [0xa1, 0x63, 0x66, 0x6f, 0x6f, 0x63, 0x62, 0x61, 0x72],
+        ".cbor",
+    );
+    let tmpl = file_with_contents(r#"Hello {{ foo }}!"#);
+
+    assert_cmd_snapshot!(
+        cli()
+            .arg(tmpl.path())
+            .arg(input.path()),
+        @r###"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Hello bar!
+
+    ----- stderr -----
+    "###);
+}
+
+#[test]
 #[cfg(feature = "ini")]
 fn test_ini() {
     let input = file_with_contents_and_ext("[section]\nfoo = bar", ".ini");
@@ -265,6 +288,128 @@ fn test_ini() {
     exit_code: 0
     ----- stdout -----
     Hello bar!
+
+    ----- stderr -----
+    "###);
+}
+
+#[test]
+#[cfg(feature = "preserve_order")]
+fn test_preserve_order_json() {
+    let input = file_with_contents_and_ext(r#"{"x": {"c": 3, "a": 1, "b": 2}}"#, ".json");
+    let tmpl =
+        file_with_contents("{% for key, value in x|items %}{{ key }}: {{ value }}\n{% endfor %}");
+
+    assert_cmd_snapshot!(
+        cli()
+            .arg(tmpl.path())
+            .arg(input.path()),
+        @r###"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    c: 3
+    a: 1
+    b: 2
+
+
+    ----- stderr -----
+    "###);
+}
+
+#[test]
+#[cfg(all(feature = "preserve_order", feature = "yaml"))]
+fn test_preserve_order_yaml() {
+    let input = file_with_contents_and_ext(
+        r#"
+x:
+  c: 3
+  a: 1
+  b: 2
+"#,
+        ".yaml",
+    );
+    let tmpl =
+        file_with_contents("{% for key, value in x|items %}{{ key }}: {{ value }}\n{% endfor %}");
+
+    assert_cmd_snapshot!(
+        cli()
+            .arg(tmpl.path())
+            .arg(input.path()),
+        @r###"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    c: 3
+    a: 1
+    b: 2
+
+
+    ----- stderr -----
+    "###);
+}
+
+#[test]
+#[cfg(all(feature = "preserve_order", feature = "toml"))]
+fn test_preserve_order_toml() {
+    let input = file_with_contents_and_ext(
+        r#"
+[x]
+c = 3
+a = 1
+b = 2
+"#,
+        ".toml",
+    );
+    let tmpl =
+        file_with_contents("{% for key, value in x|items %}{{ key }}: {{ value }}\n{% endfor %}");
+
+    assert_cmd_snapshot!(
+        cli()
+            .arg(tmpl.path())
+            .arg(input.path()),
+        @r###"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    c: 3
+    a: 1
+    b: 2
+
+
+    ----- stderr -----
+    "###);
+}
+
+#[test]
+#[cfg(all(feature = "preserve_order", feature = "cbor"))]
+fn test_preserve_order_cbor() {
+    let input = file_with_contents_and_ext(
+        [
+            0xa1, // map(1)
+            0x61, 0x78, // "x"
+            0xa3, // map(3)
+            0x61, 0x63, 0x03, // "c": 3
+            0x61, 0x61, 0x01, // "a": 1
+            0x61, 0x62, 0x02, // "b": 2
+        ],
+        ".cbor",
+    );
+    let tmpl =
+        file_with_contents("{% for key, value in x|items %}{{ key }}: {{ value }}\n{% endfor %}");
+
+    assert_cmd_snapshot!(
+        cli()
+            .arg(tmpl.path())
+            .arg(input.path()),
+        @r###"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    c: 3
+    a: 1
+    b: 2
+
 
     ----- stderr -----
     "###);
@@ -492,4 +637,120 @@ fn test_line_statement() {
 
     ----- stderr -----
     "###);
+}
+
+#[test]
+#[allow(clippy::suspicious_command_arg_space)]
+fn test_template_string() {
+    assert_cmd_snapshot!(
+        cli()
+            .arg("-tHello {{ name }}")
+            .arg("-Dname=Peter")
+            .arg("--no-newline"),
+        @r###"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Hello Peter
+    ----- stderr -----
+    "###);
+}
+
+#[test]
+fn test_print_config_fully_loaded() {
+    assert_cmd_snapshot!(
+        cli()
+            .arg("--strict")
+            .arg("--trim-blocks")
+            .arg("-Dvar1=value1")
+            .arg("-Dvar2=value2")
+            .arg("-Dvar3:=42")
+            .arg("-Dvar4:=true")
+            .arg("-Dvar5:=[1,2,true]")
+            .arg("--print-config"),
+        @r###"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    format = "auto"
+    autoescape = "auto"
+    include = true
+    newline = true
+    trim-blocks = true
+    lstrip-blocks = false
+    py-compat = false
+    env = false
+    strict = true
+    safe-paths = []
+    expr-out = "print"
+    fuel = 0
+
+    [syntax]
+    block-start = "{%"
+    block-end = "%}"
+    variable-start = "{{"
+    variable-end = "}}"
+    comment-start = "{#"
+    comment-end = "#}"
+    line-statement-prefix = ""
+    line-comment-prefix = ""
+
+    [defines]
+    var1 = "value1"
+    var2 = "value2"
+    var3 = 42
+    var4 = true
+    var5 = [
+        1,
+        2,
+        true,
+    ]
+
+
+    ----- stderr -----
+    "###);
+}
+
+#[test]
+fn test_load_config() {
+    let config = file_with_contents_and_ext(
+        r#"
+    [defines]
+    greeting = "Hello"
+    punctuation = "!"
+    "#,
+        ".toml",
+    );
+
+    let input = file_with_contents_and_ext(r#"{"name": "World"}"#, ".json");
+
+    assert_cmd_snapshot!(
+        cli()
+            .arg("--config-file")
+            .arg(config.path())
+            .arg("-")
+            .arg(input.path())
+            .pass_stdin("{{ greeting }} {{ name }}{{ punctuation }}"),
+        @r###"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Hello World!
+
+    ----- stderr -----
+    "###);
+}
+
+#[test]
+#[cfg(all(
+    feature = "cbor",
+    feature = "ini",
+    feature = "json5",
+    feature = "querystring",
+    feature = "toml",
+    feature = "yaml",
+))]
+fn test_help() {
+    assert_cmd_snapshot!("short_help", cli().arg("--help"));
+    assert_cmd_snapshot!("long_help", cli().arg("--long-help"));
 }